requirements.txt
fastapi
uvicorn
torch
torchvision
torchaudio
numpy
pillow
python-multipart
